Deleted Main. Renamed sample package to bestcharactersheet. Change structure to not read from dump .fxml files. Controller is not implemented yet.
